Move datatype mgmt to separate file
